Decompose data from two sensors

Decompose data from two sensors

Usage

decompose_2_fct(
  data,
  sensor1,
  sensor2,
  hour_x,
  direction1,
  direction2,
  mobility,
  norm,
  vacations = NULL,
  public_holidays = NULL,
  sensor_names = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14"),
  sensor_ids = c(9000002156, 9000001906, 9000001618, 9000003090, 9000002453, 9000001844,
    9000001877, 9000002666, 9000002181, 9000002707, 9000003703, 9000003746, 9000003775,
    9000003736)
)

Arguments

data

Data frame containing the data

sensor1

The segment ID for sensor 1

sensor2

The segment ID for sensor 2

hour_x

The hour to filter on

direction1

The direction for sensor 1

direction2

The direction for sensor 2

mobility

A character vector indicating the mobility types to include in the filter

norm

A character indicating whether to normalize the results

vacations

vacations period for filtering function

public_holidays

public holidays period for filtering function

sensor_names

A character vector containing the name of each sensor that is displayed to the user

sensor_ids

A character vector containing the identifier name for each vector

Value

A list containing the decomposed data, seasonal data, peaks, and correlation