Filter by selected criteria and aggregating traffics.

Not all criteria need to be filled in. Unfilled criteria are set by default so that no filtering is performed.

Usage

filter_agg(
  data,
  date_range = NULL,
  segments = NULL,
  direction = NULL,
  modes = NULL,
  weekdays = NULL,
  hours = NULL,
  uptime_quality = TRUE
)

Arguments

data

Traffic Data Frame

date_range

Date vector, c("aaaa-mm-jj","aaaa-mm-jj")

segments

Vector of character. Ids of desired segments.

direction

Vector of character. Direction of the street (lft, right, both).

modes

Vector of character. Type(s) of mobility: c("car","heavy","pedestrian","bike")

weekdays

Vector of character. Weekday(s) choosen.

hours

Integer vector. Hours choosen, default to the all day.

Value

the filtered data, molten by mode and direction, with new columns :

mode, the mode, one row per different transportation mode + date/hour + direction

direction, the direction, one row per different transportation mode + date/hour + direction

traffic_sum, the traffic for this mode/direction on this specific date/hour

Examples

date_range = as.Date(c('2022-01-01','2022-01-08'))
filter_agg(traffic,
  date_range = date_range,
  segments = 'RteVitre-06',
  direction = 'lft',
  modes = 'pedestrian',
  weekdays = 'saturday',
  hours = 12:14,
  uptime_quality = TRUE
  )
#> $data
#>                  date        day hour  weekday holiday         vacation
#> 1 2022-01-01 12:00:00 2022-01-01   12 saturday    TRUE Vacances de Noël
#> 2 2022-01-01 13:00:00 2022-01-01   13 saturday    TRUE Vacances de Noël
#> 3 2022-01-08 12:00:00 2022-01-08   12 saturday   FALSE      No vacation
#> 4 2022-01-08 13:00:00 2022-01-08   13 saturday   FALSE      No vacation
#> 5 2022-01-08 14:00:00 2022-01-08   14 saturday   FALSE      No vacation
#>   segment_name  v85
#> 1  RteVitre-06  6.5
#> 2  RteVitre-06 40.0
#> 3  RteVitre-06 43.0
#> 4  RteVitre-06 44.5
#> 5  RteVitre-06 44.0
#>                                                                   car_speed_hist_0to70plus
#> 1                                                                 100, 0, 0, 0, 0, 0, 0, 0
#> 2        0.000000, 2.222222, 11.111111, 63.333333, 22.222222, 0.000000, 0.000000, 1.111111
#> 3 6.3333333, 3.6666667, 3.6666667, 41.3333333, 38.6666667, 6.0000000, 0.0000000, 0.3333333
#> 4 1.7667845, 3.1802120, 3.5335689, 41.6961131, 45.2296820, 4.2402827, 0.3533569, 0.0000000
#> 5 0.9740260, 0.3246753, 2.9220779, 48.7012987, 40.9090909, 5.8441558, 0.0000000, 0.3246753
#>                                                                                                                                                                                                                                                               car_speed_hist_0to120plus
#> 1                                                                                                                                                                                                           0, 100, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 2                           0.000000, 0.000000, 0.000000, 2.222222, 3.333333, 7.777778, 22.222222, 41.111111, 16.666667, 5.555556, 0.000000, 0.000000, 0.000000, 0.000000, 0.000000, 0.000000, 1.111111, 0.000000, 0.000000, 0.000000, 0.000000, 0.000000, 0.000000, 0.000000, 0.000000
#> 3  3.6666667, 2.6666667, 1.6666667, 2.0000000, 1.0000000, 2.6666667, 14.0000000, 27.3333333, 29.0000000, 9.6666667, 5.3333333, 0.6666667, 0.0000000, 0.0000000, 0.3333333, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000
#> 4 1.0600707, 0.7067138, 1.4134276, 1.7667845, 0.7067138, 2.8268551, 13.7809187, 27.9151943, 28.9752650, 16.2544170, 3.5335689, 0.7067138, 0.0000000, 0.3533569, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000
#> 5 0.9740260, 0.0000000, 0.3246753, 0.0000000, 0.9740260, 1.9480519, 14.2857143, 34.4155844, 28.8961039, 12.0129870, 4.8701299, 0.9740260, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.3246753, 0.0000000, 0.0000000, 0.0000000, 0.0000000, 0.0000000
#>      uptime uptime_quality       mode direction traffic_sum
#> 1 0.7875000           TRUE pedestrian       lft   16.507937
#> 2 0.7277778           TRUE pedestrian       lft   10.992366
#> 3 0.7808333           TRUE pedestrian       lft    0.000000
#> 4 0.7922222           TRUE pedestrian       lft    1.262272
#> 5 0.7922222           TRUE pedestrian       lft    1.262272
#> 
#> $segment
#> [1] "RteVitre-06"
#> 
#> $mode
#> [1] "pedestrian"
#> 
#> $direction
#> [1] "lft"
#> 
#> $weekday
#> [1] "saturday"
#> 
#> $hour
#> [1] 12 13 14
#>