gg_na_heatmap This function generates a heatmap representation of the proportion of missing values (NA) for different sensors over monthly periods during a selected time frame.

gg_na_heatmap This function generates a heatmap representation of the proportion of missing values (NA) for different sensors over monthly periods during a selected time frame.

Usage

gg_na_heatmap(
  data,
  start = NULL,
  end = NULL,
  hours = "ALL",
  list_weekday = "ALL",
  pub_holidays = "YES",
  holidays = "YES",
  sensor_ids = c(9000002156, 9000001906, 9000001618, 9000003090, 9000002453, 9000001844,
    9000001877, 9000002666, 9000002181, 9000002707, 9000003703, 9000003746, 9000003775,
    9000003736, 9000004971, 9000004130, 9000004042, 9000004697),
  sensor_names = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14", "RueManoirs-15", "RueToursCarree-16",
    "PlaceHotelDeVille-17", "BoulevardLiberte-18")
)

Arguments

data

Data frame. A data frame containing road traffic data.

start

Character. The start date in the format "yyyy-mm-dd".

end

Character. The end date in the format "yyyy-mm-dd".

hours

Vector. A vector containing hours on which you want to see the representation, e.g., c("05h", "18h").

list_weekday

Vector. A vector containing weekdays in French, e.g., c("Monday", "Sunday").

pub_holidays

Character. "YES" if you want to see results with public holidays included; "NO" if you want to exclude them from the representation; "ONLY" if you only want to see them in the result.

holidays

Character. "YES" if you want to see results including holidays; "NO" if you want to exclude them from the representation; "ONLY" if you only want to see them in the result.

sensor_ids

list with the ids of all the studied sensors

sensor_names

list with the name of all the studied sensors

Value

A ggplot graph representing the heatmap of NA proportions for different sensors over monthly periods.