Impute missing data for traffic analysis

This function imputes missing data for traffic analysis using Random Forest models.

Usage

impute_missing_data(
  data,
  sensors_id = NULL,
  transport_type = "vehicle",
  threshold_uptime = 0.5,
  add_vars = NULL,
  min.node.size = NULL,
  mtry = NULL,
  num_trees = 500
)

Arguments

data

A data frame containing the traffic data.

sensors_id

Character vector. Id of the sensors to include in the analysis. Default is NULL (all sensors).

transport_type

Character. Type of transport to impute. Options are "car", "vehicle", "heavy", or "all". Default is "vehicle".

threshold_uptime

Numeric. Threshold for uptime to determine missing values. Default is 0.5.

add_vars

Character vector. Additional variables used for prediction. Default is NULL.

min.node.size

Integer or NULL. Minimum node size for the Random Forest model. If NULL, the default value of 5 for regression is used.

mtry

Integer or NULL. Number of variables randomly sampled as candidates at each split. If NULL, the default value (number of variables / 3 for regression) is used.

num_trees

Integer. Number of trees to grow in the Random Forest model. Default is 500.

Value

A data frame with imputed values for the specified transport type and a new column indicating whether the values were imputed or original.

Details

The function requires specific columns to be present in the input data: day_of_month, hour, weekday, month, year, vacation, week_number, segment_id, date, car, heavy, and uptime.

Vehicle is calculated as the sum of cars and heavy vehicles providing a good estimate of the total traffic.

When transport_type is "all", the function imputes values for car, vehicle, and heavy separately.

The function uses the uptime column to determine which data points need imputation. Values with uptime below the threshold_uptime are considered missing and are imputed.

Time interval could be hourly or quarterly. The function will automatically detect the time interval based on the data and add a minute variable if it's quarterly.

Examples

if (FALSE) {
 traffic_clean <- retrieve_missing_data(traffic,remove_data = TRUE,successive_day = 2)
 traffic_imputed <- impute_missing_data(traffic,
   sensors_id = 9000001844,
   transport_type = "vehicle",
   threshold_uptime = 0.5)
 table(traffic_imputed$imputed)
}