Write or update the sensor data in the data folder

Writes or updates the sensor data in the data folder. It retrieves the data for the specified sensor between date1 and date2 (inclusive) using the retrieve_sensor function, and then converts certain columns to character strings before writing the data to a RData file in the data folder.

Usage

write_update_data(
  id_sensor,
  date1,
  date2,
  sensor_names = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14", "RueManoirs-15", "RueToursCarree-16",
    "PlaceHotelDeVille-17", "BoulevardLiberte-18"),
  sensor_ids = c(9000002156, 9000001906, 9000001618, 9000003090, 9000002453, 9000001844,
    9000001877, 9000002666, 9000002181, 9000002707, 9000003703, 9000003746, 9000003775,
    9000003736, 9000004971, 9000004130, 9000004042, 9000004697)
)

Arguments

id_sensor

Numeric. ID of the sensor

date1

Date. Start date "aaaa-mm-jj"

date2

Date. End date "aaaa-mm-jj"

sensor_names

list with the name of all the studied sensors

sensor_ids

list with the ids of all the studied sensors